fixed the compilation error on x86_64 reported by J.

+++ src/ngx_http_chunkin_filter_module.h
@@ -22,7 +22,7 @@ typedef struct {
     size_t              chunks_written_size;
     int                 chunks_count;
 
-    size_t              raw_body_size;
+    off_t               raw_body_size;
 
     ngx_chain_t         *chunks;
     ngx_chain_t         **next_chunk;
